Update to avoid compiler warning.

+++ src/humlib.cpp
@@ -1,7 +1,7 @@
 //
 // Programmer:    Craig Stuart Sapp <craig@ccrma.stanford.edu>
 // Creation Date: Sat Aug  8 12:24:49 PDT 2015
-// Last Modified: Fri Aug  9 22:21:56 EDT 2019
+// Last Modified: Fri Aug  9 22:27:23 EDT 2019
 // Filename:      /include/humlib.cpp
 // URL:           https://github.com/craigsapp/humlib/blob/master/src/humlib.cpp
 // Syntax:        C++11
@@ -43944,7 +43944,7 @@ void Tool_homophonic::processFile(HumdrumFile& infile) {
 	}
 
 	// change Y N Y patterns to Y Y Y
-	for (int i=1; i<data.size() - 1; i++) {
+	for (int i=1; i<(int)data.size() - 1; i++) {
 		if (m_homophonic[data[i]] == "Y") {
 			continue;
 		}
